refactor(gateway): one home for the envelope serializers; barrels say what is public
The two reviews on this branch. The request and response serializers stood
in two copies inside the new package, and they belong beside the readers that
take the same envelopes back in, so they move into protocol.ts with the
refusal the protocol itself answers.

The barrels said more than they meant. @sidecar/gateway is a package whose
every name is part of the contract, so it is one door per module, which is
what packages/AGENTS.md says such a barrel is; @sidecar/host is not, so its
barrel is the hand-listed set a client actually reaches, pruned of the
nineteen names nothing outside the package named. The node ids' values now
say what their names do. The desktop's drain was a two-line function with
one caller, folded into it, and the prose inside the host that still called
the desktop the composer says the host.
