+++ packages/host/src/index.ts
@@ -8,27 +8,14 @@
  * its own process today and reaches it over the in-process transport; a
  * process on the other side of a socket is the same host over another
  * transport.
+ *
+ * This package's modules are not all public, which is what the hand-listed
+ * barrel says: every name here has a client that reaches it, and a name a
+ * client does not reach yet is added when one arrives.
  */
-export { arrivalBeatOwed, countsFirstAnnouncement } from "./arrival-flow.js";
-export {
-  type BrainPublicationAgent,
-  type BrainPublicationDependencies,
-  followBrainRequests,
-  publishAsk,
-  publishRuns,
-  REJECTED_SUBMISSION,
-} from "./brain/publication.js";
-export { calendarOnboardingOwed } from "./calendar-onboarding-flow.js";
-export {
-  composeHost,
-  type Host,
-  type HostSeams,
-  RECEIVER_REPORT_KIND,
-} from "./compose-host.js";
-export {
-  type ConversationOperations,
-  conversationOperations,
-} from "./conversation-operations.js";
+export { REJECTED_SUBMISSION } from "./brain/publication.js";
+export { composeHost, type Host, type HostSeams, RECEIVER_REPORT_KIND } from "./compose-host.js";
+export type { ConversationOperations } from "./conversation-operations.js";
 export {
   INTRODUCTION_FADE_MS,
   INTRODUCTION_HANDOFF_READY_MS,
@@ -36,36 +23,17 @@ export {
   INTRODUCTION_RENDER_DEADLINE_MS,
   shouldRunIntroduction,
 } from "./introduction-flow.js";
-export { type JsonStateFile, jsonStateFile } from "./json-state-file.js";
+export { jsonStateFile } from "./json-state-file.js";
 export {
   HOST_NATIVE_NODE_ID,
   HOST_NODE_CAPABILITY,
   HOST_NODE_CAPABILITY_LIST,
   HOST_OPERATOR_CLIENT_ID,
-  type HostNodeCapability,
 } from "./node-capabilities.js";
-export {
-  type OnboardingState,
-  onboardingStateFile,
-} from "./onboarding-state.js";
-export {
-  createGatewayOperator,
-  type GatewayHistoryChange,
-  type GatewayOperator,
-} from "./operator.js";
+export { onboardingStateFile } from "./onboarding-state.js";
+export { createGatewayOperator, type GatewayOperator } from "./operator.js";
 export { type RunMode, runModeFor, sentryReportingEnabled } from "./run-mode.js";
-export {
-  createGatewayService,
-  type GatewayService,
-  type GrantedWords,
-} from "./service.js";
-export {
-  createSessionActPerformer,
-  NodeAnswerLostError,
-  OPEN_REFUSAL,
-  type SessionActPerformer,
-  type SessionActPerformerDependencies,
-} from "./session-act-performer.js";
-export { type SecretCipher, SettingsStore } from "./settings-store.js";
+export { createGatewayService, type GrantedWords } from "./service.js";
+export { OPEN_REFUSAL, type SessionActPerformer } from "./session-act-performer.js";
 export { runtimeStoreWorkerPath } from "./store-path.js";
 export { VoiceReceiver } from "./voice-receiver.js";
